Move a few more helpers out of libs/backend and into libs/utils
