+++ libs/backend/src/cast_vote_records/build_cast_vote_record.ts
@@ -27,7 +27,11 @@ import {
   YesNoContest,
   YesNoVote,
 } from '@votingworks/types';
-import { getContestsForBallotPage, getMarkStatus } from '@votingworks/utils';
+import {
+  getContestsForBallotPage,
+  getMarkStatus,
+  getWriteInCount,
+} from '@votingworks/utils';
 
 import {
   ContestOptionPositionMap,
@@ -403,41 +407,6 @@ function buildOriginalSnapshot({
       })
     ),
   };
-}
-
-/**
- * Determines the number of write-in candidates in a {@link VotesDict}
- */
-export function getWriteInCount(votes: VotesDict): number {
-  let count = 0;
-  for (const vote of Object.values(votes)) {
-    if (vote) {
-      for (const voteOption of vote) {
-        if (typeof voteOption !== 'string' && voteOption.isWriteIn) {
-          count += 1;
-        }
-      }
-    }
-  }
-
-  return count;
-}
-
-/**
- * Determines whether a {@link VotesDict} contains any write-in candidates
- */
-export function hasWriteIns(votes: VotesDict): boolean {
-  for (const vote of Object.values(votes)) {
-    if (vote) {
-      for (const voteOption of vote) {
-        if (typeof voteOption !== 'string' && voteOption.isWriteIn) {
-          return true;
-        }
-      }
-    }
-  }
-
-  return false;
 }
 
 /**
