+++ libs/utils/src/votes.ts
@@ -161,3 +161,38 @@ export function convertMarksToVotesDict(
   }
   return votesDict;
 }
+
+/**
+ * Determines the number of write-in candidates in a {@link VotesDict}
+ */
+export function getWriteInCount(votes: VotesDict): number {
+  let count = 0;
+  for (const vote of Object.values(votes)) {
+    if (vote) {
+      for (const voteOption of vote) {
+        if (typeof voteOption !== 'string' && voteOption.isWriteIn) {
+          count += 1;
+        }
+      }
+    }
+  }
+
+  return count;
+}
+
+/**
+ * Determines whether a {@link VotesDict} contains any write-in candidates
+ */
+export function hasWriteIns(votes: VotesDict): boolean {
+  for (const vote of Object.values(votes)) {
+    if (vote) {
+      for (const voteOption of vote) {
+        if (typeof voteOption !== 'string' && voteOption.isWriteIn) {
+          return true;
+        }
+      }
+    }
+  }
+
+  return false;
+}
